# Drastically improve performance when resolving large dossiers.